Rename ParMap tests as ParZip (#2821)
