+++ arrow-libs/fx/arrow-fx-coroutines/src/commonTest/kotlin/arrow/fx/coroutines/parZip/ParZip3Test.kt
@@ -1,4 +1,4 @@
-package arrow.fx.coroutines.parMapN
+package arrow.fx.coroutines.parZip
 
 import arrow.core.Either
 import arrow.fx.coroutines.Atomic
@@ -22,8 +22,8 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.CoroutineScope
 
-class ParMap3Test : StringSpec({
-    "parMapN 3 runs in parallel" {
+class ParZip3Test : StringSpec({
+    "parZip 3 runs in parallel" {
       checkAll(Arb.int(), Arb.int(), Arb.int()) { a, b, c ->
         val r = Atomic("")
         val modifyGate1 = CompletableDeferred<Unit>()
@@ -51,7 +51,7 @@ class ParMap3Test : StringSpec({
       }
     }
 
-    "Cancelling parMapN 3 cancels all participants" {
+    "Cancelling parZip 3 cancels all participants" {
         val s = Channel<Unit>()
         val pa = CompletableDeferred<ExitCase>()
         val pb = CompletableDeferred<ExitCase>()
@@ -76,7 +76,7 @@ class ParMap3Test : StringSpec({
         pc.await().shouldBeTypeOf<ExitCase.Cancelled>()
     }
 
-    "parMapN 3 cancels losers if a failure occurs in one of the tasks" {
+    "parZip 3 cancels losers if a failure occurs in one of the tasks" {
       checkAll(
         Arb.throwable(),
         Arb.element(listOf(1, 2, 3)),
@@ -105,7 +105,7 @@ class ParMap3Test : StringSpec({
       }
     }
 
-    "parMapN CancellationException on right can cancel rest" {
+    "parZip CancellationException on right can cancel rest" {
       checkAll(Arb.string(), Arb.int(1..3)) { msg, cancel ->
         val s = Channel<Unit>()
         val pa = CompletableDeferred<ExitCase>()
