+++ arrow-libs/fx/arrow-fx-coroutines/src/commonTest/kotlin/arrow/fx/coroutines/parZip/ParZip5Test.kt
@@ -1,4 +1,4 @@
-package arrow.fx.coroutines.parMapN
+package arrow.fx.coroutines.parZip
 
 import arrow.core.Either
 import arrow.core.Tuple5
@@ -23,8 +23,8 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.CoroutineScope
 
-class ParMap5Test : StringSpec({
-    "parMapN 5 runs in parallel" {
+class ParZip5Test : StringSpec({
+    "parZip 5 runs in parallel" {
       checkAll(Arb.int(), Arb.int(), Arb.int(), Arb.int(), Arb.int()) { a, b, c, d, e ->
         val r = Atomic("")
         val modifyGate1 = CompletableDeferred<Unit>()
@@ -64,7 +64,7 @@ class ParMap5Test : StringSpec({
       }
     }
 
-    "Cancelling parMapN 5 cancels all participants" {
+    "Cancelling parZip 5 cancels all participants" {
         val s = Channel<Unit>()
         val pa = CompletableDeferred<ExitCase>()
         val pb = CompletableDeferred<ExitCase>()
@@ -100,7 +100,7 @@ class ParMap5Test : StringSpec({
         pe.await().shouldBeTypeOf<ExitCase.Cancelled>()
     }
 
-    "parMapN 5 cancels losers if a failure occurs in one of the tasks" {
+    "parZip 5 cancels losers if a failure occurs in one of the tasks" {
       checkAll(
         Arb.throwable(),
         Arb.element(listOf(1, 2, 3, 4, 5)),
@@ -135,7 +135,7 @@ class ParMap5Test : StringSpec({
       }
     }
 
-    "parMapN CancellationException on right can cancel rest" {
+    "parZip CancellationException on right can cancel rest" {
       checkAll(Arb.string(), Arb.int(1..5)) { msg, cancel ->
         val s = Channel<Unit>()
         val pa = CompletableDeferred<ExitCase>()
