+++ arrow-libs/fx/arrow-fx-coroutines/src/commonTest/kotlin/arrow/fx/coroutines/parZip/ParZip6Test.kt
@@ -1,4 +1,4 @@
-package arrow.fx.coroutines.parMapN
+package arrow.fx.coroutines.parZip
 
 import arrow.core.Either
 import arrow.core.Tuple6
@@ -23,8 +23,8 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.CoroutineScope
 
-class ParMap6Test : StringSpec({
-    "parMapN 6 runs in parallel" {
+class ParZip6Test : StringSpec({
+    "parZip 6 runs in parallel" {
       checkAll(Arb.int(), Arb.int(), Arb.int(), Arb.int(), Arb.int(), Arb.int()) { a, b, c, d, e, f ->
         val r = Atomic("")
         val modifyGate1 = CompletableDeferred<Unit>()
@@ -70,7 +70,7 @@ class ParMap6Test : StringSpec({
       }
     }
 
-    "Cancelling parMapN 6 cancels all participants" {
+    "Cancelling parZip 6 cancels all participants" {
         val s = Channel<Unit>()
         val pa = CompletableDeferred<ExitCase>()
         val pb = CompletableDeferred<ExitCase>()
@@ -103,7 +103,7 @@ class ParMap6Test : StringSpec({
         pf.await().shouldBeTypeOf<ExitCase.Cancelled>()
     }
 
-    "parMapN 6 cancels losers if a failure occurs in one of the tasks" {
+    "parZip 6 cancels losers if a failure occurs in one of the tasks" {
       checkAll(
         Arb.throwable(),
         Arb.element(listOf(1, 2, 3, 4, 5, 6))
@@ -142,7 +142,7 @@ class ParMap6Test : StringSpec({
       }
     }
 
-    "parMapN CancellationException on right can cancel rest" {
+    "parZip CancellationException on right can cancel rest" {
       checkAll(Arb.string(), Arb.int(1..6)) { msg, cancel ->
         val s = Channel<Unit>()
         val pa = CompletableDeferred<ExitCase>()
